♻️refactor: Spring Data Jpa의 공식 COUNT의 반환 타입은 Long이기 때문에 그에 맞게 수정

+++ src/main/java/novaminds/gradproj/domain/recipe/repository/RecipeCommentRepository.java
@@ -20,7 +20,7 @@ public interface RecipeCommentRepository extends JpaRepository<RecipeComment, Lo
     @Query("SELECT COUNT(c) " +
             "FROM RecipeComment c " +
             "WHERE c.recipe.id = :recipeId")
-    int countCommentsByRecipeId(@Param("recipeId") Long recipeId);
+    Long countCommentsByRecipeId(@Param("recipeId") Long recipeId);
 
     //댓글 ID들로 작성자 ID만 조회
     @Query(value = "SELECT author_id " +
